[Model] Adding User Name Feature on Home Page
Added a feature to display the user's name on the home page.
First, defined the Home class, which is a child of the Controller class. Inside the Home class, there is an index() method used to render the home page.
Sent the title and user's name data to the view to be displayed on the browser tab.
The user's name data is obtained from the getUser() method in the User_model class.
Next, in the views\home\index.php file, displayed the user's name sent from the controller.